fix: Windows cross-platform compatibility for PyPI users
Guard POSIX-only includes with _WIN32, add compat.h shims (usleep/sleep/getpid/gettimeofday),
fix time_t FFI typedef (long→long long for Win64 ABI), bundle MinGW runtimes via delvewheel,
add macOS delocate repair, add x86_64 macOS arch, and document MinGW as build-time-only dep.

+++ python/src/gigavector/_ffi.py
@@ -16,7 +16,7 @@ ffi: FFIType = FFI()
 # Keep in sync with include/gigavector/gigavector.h
 ffi.cdef(
     """
-typedef long time_t;  // Define time_t for FFI
+typedef long long time_t;
 typedef enum { GV_INDEX_TYPE_KDTREE = 0, GV_INDEX_TYPE_HNSW = 1, GV_INDEX_TYPE_IVFPQ = 2, GV_INDEX_TYPE_SPARSE = 3, GV_INDEX_TYPE_FLAT = 4, GV_INDEX_TYPE_IVFFLAT = 5, GV_INDEX_TYPE_PQ = 6, GV_INDEX_TYPE_LSH = 7 } GV_IndexType;
 typedef enum { GV_DISTANCE_EUCLIDEAN = 0, GV_DISTANCE_COSINE = 1, GV_DISTANCE_DOT_PRODUCT = 2, GV_DISTANCE_MANHATTAN = 3, GV_DISTANCE_HAMMING = 4 } GV_DistanceType;
